pyproject.toml
# Purpose: Define the Python package metadata, runtime dependencies, and local developer tooling.
# Input/Output: Packaging and lint/test tooling read this file to install and validate the project.
# Invariants: Python 3.11+ is required because the services rely on modern typing and FastAPI patterns.
# Debugging: If install or import errors occur, start by checking dependency ranges and package discovery here.

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "feberdin-agent-team"
version = "0.1.0"
description = "Local-first multi-agent coding system for Unraid staging workflows."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Feberdin" }]
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "jinja2>=3.1,<4.0",
  "langgraph>=0.2,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "PyYAML>=6.0,<7.0",
  "python-slugify>=8.0,<9.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11,<2.0",
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "ruff>=0.6,<1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["services*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 140
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["tests/"]
